Moved Vc/x86 switched functions from CoreFunc to MathHelper to match o2d
