+++ src/MathHelper.h
@@ -7,8 +7,166 @@
 
 #pragma once
 
+#ifdef _WIN32
+#define __restrict__ __restrict
+#endif
+
+#ifdef USE_VC
+#include <Vc/Vc>
+#endif
+
 #include <cmath>
 #include <array>
+
+// helper functions: exp, recip, sqrt, rsqrt, oor1p5
+
+#ifdef USE_VC
+template <class S>
+static inline S my_exp(const S _in) {
+  return Vc::exp(_in);
+}
+template <>
+inline float my_exp(const float _in) {
+  return std::exp(_in);
+}
+template <>
+inline double my_exp(const double _in) {
+  return std::exp(_in);
+}
+#else
+template <class S>
+static inline S my_exp(const S _in) {
+  return std::exp(_in);
+}
+#endif
+
+#ifdef USE_VC
+template <class S>
+static inline S my_recip(const S _in) {
+  return Vc::reciprocal(_in);
+}
+template <>
+inline float my_recip(const float _in) {
+  return 1.0f / _in;
+}
+template <>
+inline double my_recip(const double _in) {
+  return 1.0 / _in;
+}
+#else
+template <class S>
+static inline S my_recip(const S _in) {
+  return S(1.0) / _in;
+}
+#endif
+
+#ifdef USE_VC
+template <class S>
+static inline S my_sqrt(const S _in) {
+  return Vc::sqrt(_in);
+}
+template <>
+inline float my_sqrt(const float _in) {
+  return std::sqrt(_in);
+}
+template <>
+inline double my_sqrt(const double _in) {
+  return std::sqrt(_in);
+}
+#else
+template <class S>
+static inline S my_sqrt(const S _in) {
+  return std::sqrt(_in);
+}
+#endif
+
+#ifdef USE_VC
+template <class S>
+static inline S my_rsqrt(const S _in) {
+  return Vc::rsqrt(_in);
+}
+template <>
+inline float my_rsqrt(const float _in) {
+  return 1.0f / std::sqrt(_in);
+}
+template <>
+inline double my_rsqrt(const double _in) {
+  return 1.0 / std::sqrt(_in);
+}
+#else
+template <class S>
+static inline S my_rsqrt(const S _in) {
+  return S(1.0) / std::sqrt(_in);
+}
+#endif
+
+#ifdef USE_VC
+template <class S>
+static inline S oor2p5(const S _in) {
+  //return Vc::reciprocal(_in*_in*Vc::sqrt(_in));	// 234 GFlop/s
+  return Vc::rsqrt(_in) * Vc::reciprocal(_in*_in);	// 269 GFlop/s
+}
+template <>
+inline float oor2p5(const float _in) {
+  return 1.0f / (_in*_in*std::sqrt(_in));
+}
+template <>
+inline double oor2p5(const double _in) {
+  return 1.0 / (_in*_in*std::sqrt(_in));
+}
+#else
+template <class S>
+static inline S oor2p5(const S _in) {
+  return S(1.0) / (_in*_in*std::sqrt(_in));
+}
+#endif
+
+#ifdef USE_VC
+template <class S>
+static inline S oor1p5(const S _in) {
+  //return Vc::reciprocal(_in*Vc::sqrt(_in));		// 243 GFlop/s
+  return Vc::rsqrt(_in) * Vc::reciprocal(_in);		// 302 GFlop/s
+}
+template <>
+inline float oor1p5(const float _in) {
+  return 1.0f / (_in*std::sqrt(_in));
+}
+template <>
+inline double oor1p5(const double _in) {
+  return 1.0 / (_in*std::sqrt(_in));
+}
+#else
+template <class S>
+static inline S oor1p5(const S _in) {
+  return S(1.0) / (_in*std::sqrt(_in));
+}
+#endif
+
+#ifdef USE_VC
+template <class S>
+static inline S oor0p75(const S _in) {
+  const S rsqd = Vc::rsqrt(_in);
+  //return rsqd*Vc::sqrt(rsqd);				// 265 GFlop/s
+  return rsqd*rsqd*Vc::rsqrt(rsqd);			// 301 GFlop/s
+}
+template <>
+inline float oor0p75(const float _in) {
+  const float sqd = std::sqrt(_in);
+  return 1.0f / (sqd*std::sqrt(sqd));
+}
+template <>
+inline double oor0p75(const double _in) {
+  const double sqd = std::sqrt(_in);
+  return 1.0 / (sqd*std::sqrt(sqd));
+}
+#else
+template <class S>
+static inline S oor0p75(const S _in) {
+  const S sqd = std::sqrt(_in);
+  return S(1.0) / (sqd*std::sqrt(sqd));
+}
+#endif
+
 
 // normalize a 3-vector
 template <class S>
